Fix internalとpkg内を整理
