+++ pkg/host/cpu/cpu.go
@@ -2,13 +2,22 @@ package cpu
 
 import (
 	"github.com/CharVstack/CharV-backend/domain/models"
+	"github.com/shirou/gopsutil/v3/cpu"
 )
 
-func GetInfo(getInfo models.Host) models.Cpu {
-	cpuInfo := models.Cpu{
-		Counts:  getInfo.Cpu.Counts,
-		Percent: getInfo.Cpu.Percent,
+func GetInfo() (models.Cpu, error) {
+	cpuCounts, err := cpu.Counts(true)
+	if err != nil {
+		return models.Cpu{}, err
 	}
 
-	return cpuInfo
+	cpuPercent, err := cpu.Percent(0, false)
+	if err != nil {
+		return models.Cpu{}, err
+	}
+
+	return models.Cpu{
+		Counts:  cpuCounts,
+		Percent: cpuPercent[0],
+	}, nil
 }
