+++ pkg/host/memory/memory.go
@@ -2,15 +2,19 @@ package memory
 
 import (
 	"github.com/CharVstack/CharV-backend/domain/models"
+	"github.com/shirou/gopsutil/v3/mem"
 )
 
-func GetInfo(getInfo models.Host) models.Memory {
-	memoryInfo := models.Memory{
-		Total:       getInfo.Mem.Total,
-		Used:        getInfo.Mem.Used,
-		Free:        getInfo.Mem.Free,
-		UsedPercent: getInfo.Mem.UsedPercent,
+func GetInfo() (models.Memory, error) {
+	memInfo, err := mem.VirtualMemory()
+	if err != nil {
+		return models.Memory{}, err
 	}
 
-	return memoryInfo
+	return models.Memory{
+		Total:       memInfo.Total,
+		Used:        memInfo.Used,
+		Free:        memInfo.Free,
+		UsedPercent: memInfo.UsedPercent,
+	}, nil
 }
