+++ pkg/host/storage/storage.go
@@ -1,28 +1,13 @@
 package storage
 
 import (
-	"github.com/CharVstack/CharV-backend/domain/models"
+	"github.com/shirou/gopsutil/v3/disk"
 )
 
-func GetInfo(getInfo models.Host) []models.StoragePool {
-	var storageInfoPointer = make([]*models.StoragePool, len(getInfo.StoragePools))
-
-	for getInfoIndex, pool := range getInfo.StoragePools {
-		getStoragePool := models.StoragePool{
-			Name:      pool.Name,
-			TotalSize: pool.TotalSize,
-			UsedSize:  pool.UsedSize,
-			Path:      pool.Path,
-			Status:    pool.Status,
-		}
-		storageInfoPointer[getInfoIndex] = &getStoragePool
+func GetSize(path string) (size uint64, usedSize uint64, err error) {
+	diskUsage, err := disk.Usage(path)
+	if err != nil {
+		return 0, 0, err
 	}
-
-	var storageInfo = make([]models.StoragePool, len(storageInfoPointer))
-
-	for getInfoIndex, pool := range storageInfoPointer {
-		storageInfo[getInfoIndex] = *pool
-	}
-
-	return storageInfo
+	return diskUsage.Total, diskUsage.Used, err
 }
